Python: Prevent bad `fastTC` in nested loops query
This was one instance where inlining didn't actually affect the join
order (and so the compiler decided to materialise the full `fastTC`
without using any of the local context). In this case, it seems to be
sufficient to simply manually specialise the `contains` relation to the
particular setup needed here.

(Note that the first occurrence of `contains` in this query is _not_ the
same predicate -- it's a version that works on `StmtList`s.)