Python: Prevent bad use of `contains` in `getCase`
This definition of `getCase` was inlined in `getScope` and then
evaluated in its entirety (without any magic) which turned out to be
costly. This meant that any use of `getScope` would pay this penalty
if not already present in the cache.

The rewrite in this commit _should_ have the same outwards behaviour,
since the parent of a `Pattern` is either a pattern itself, or a `Case`.
However, by restricting to just these classes, we never materialise the
entire `contains` relation.