Python: Fix `getCase` again
Turns out that in isolation, `getCase` is fine. What was killing the
performance was the fact that it got inlined into `getScope`, where it
resulted in the full evaluation of the flipped `contains` relation.

I think `nomagic` is sensible here -- I would not expect there to be so
many results for this predicate that evaluating it would become a
problem (note for instance that it already is restricted to relating
`Pattern`s and `Case`s).