Python: Rewrite sax XML tests
The tests for type-trackers were not that interesting, since they did
not have XML input in both cases, which is the problem we were trying
hard to solve.

I did keep the test-case of not-user-supplied url alive as well though
:+1:

I added OK/NOT OK annotations.

Notice that we report all 4 kinds of vulnerabilities on line 93