InlineExpectationsTest: Allow URLs as values
For example,

```py
requests.get("http://example.com") # $  clientRequestUrlPart="http://example.com"
```

would not have been allowed previously, since the `//` part of the value
for the URL would have caused the rest of the line to be treated as a
normal (non-interpreted) comment.

Requiring a space before the `//` allows this specific example, while
still support all the existing uses of `//` to break expectation tests,
which is usually to add a comment explaining why the annotation is like
it is.