Python: Add call-graph and argument-passing meta-queries
These also only makes sense in the presence of both points-to and
type-tracking based data-flow implementations, so should be deleted
before merging this PR.

Also changed the definition of a relevant call-target, so it's only what
is in the actual source code, which is what we want in the future! (so
what we're designing type-tracking to handle)

I also changed terminology from `callee` to `target`. It felt more
natural this way in my opinion.