Inline Expectation Tests: Allow `tag[foo bar]`
This is partly motivated by the MaD tests which looks much better now in
my opinion.

I also wanted this for testing argument passing. In Python we're
adopting the same argument positions as Ruby has
[here](https://github.com/github/codeql/blob/4f3751dfea06c84b97685301739f5303b9678a18/ruby/ql/lib/codeql/ruby/dataflow/internal/DataFlowDispatch.qll#L508-L540)

So it would be nice if `arg[keyword foo]=...` was allowed, without
having to transform the `toString()` result of an argument position into
something without a space.