ControlFlow: add further ordering fallback
For ordering the edges of the control flow graph we checked
lexicographically the location of the successor node followed by the
successor type. This adds also the string representation of the
successor node as a last resort in case the previous two coincide.

This should fix an issue we encountered in Swift (where we currently
have many entities with unknown locations) that was causing a test
to be flaky.