revise PhoenixSmashes.txt
Takes this Artifact Splice off of the thing hate can not defeat Beloved are head to footer and god right above her direct to master...main stage ignition. Recover what is marked and make it mine. If you cannot relate to this in anyway that is the light and dark, both are good! Dare not by questioning harshly with the 3 W's.  Devil's playground no more, changelog  I suggest, "Edens_Acute_Trigger_Many_Entities" |  (-:eat-me;-)  |