Swift: rework inter-process race strategy
An `ExclusiveFile` class is introduced, that is successfully created
only for the first process that starts processing it.

An `ExclusiveFile` is created from a working (temporary) path and a
target path. First the target file is touched with
`llvm::sys::fs::openFileForWrite` with `CD_CreateNew` (effectively
wrapping around `O_EXCL`), so that any following process trying to do
the same will be aware that someone else already started the work. Then
the actual trap content will still be written to the temporary path.
`ExclusiveFile::commit` can be called to move the temporeary content
to the target file.

Some of the file system logic contained in `SwiftExtractor.cpp` has been
moved to this class, and two TODOs are solved:
* introducing a better inter process file collision avoidance strategy
  (in particular this will save unneeded work, as previously all
  processes would extract everything, each throwing away what was
  previously extracted)
* better error handling for trap output operations: if unable to write
  to the trap file (or carry out other basic file operations), we just
  abort.