New workflow to add CodeQL PRs to projects
This workflow introduces some automation to add issues that are tagged with the `C#`, 'Go`, `Java` OR `Kotlin` labels to be added to the appropriate GitHub project.

The workflow is generically named "Add CodeQL pull requests to project" with the intent that future jobs might be added to the workflow to triage PRs in this repo to other projects.